+++ packages/rnative/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @lezzform/react-native
 
+## 0.0.44
+
+### Patch Changes
+
+- fix import and button props
+
+## 0.0.43
+
+### Patch Changes
+
+- add customizable button
+
 ## 0.0.42
 
 ### Patch Changes